refactor: 폴더 구조 main과 일치시킴
